@@ -2,28 +2,29 @@ source 'https://rubygems.org'
 
 ruby '2.0.0'
 
-gem 'rails', '3.2.13'
+gem 'rails',            '4.0.0'
+gem 'sass-rails',       '~> 4.0.0'
+gem 'uglifier',         '>= 1.3.0'
+gem 'coffee-rails',     '~> 4.0.0'
 gem 'jquery-rails'
-gem 'devise'
-gem 'simple_form'
+
+gem 'puma'
+
+gem 'devise',           '3.0.0.rc'
+gem "simple_form",      '3.0.0.rc'
+gem 'kaminari'
+
 gem "paperclip", "~> 3.0"
 gem 'aws-sdk'
-gem 'will_paginate', '3.0.3'
-gem 'bootstrap-will_paginate'
+
+gem 'anjlab-bootstrap-rails', '>= 2.3', :require => 'bootstrap-rails'
 gem 'bootstrap-wysihtml5-rails'
-gem 'thin'
 
 group :production do
 	gem "pg"
+  gem 'rails_12factor'
 end
 
 group :development, :test do
 	gem 'sqlite3'
-end
-
-group :assets do
-  gem 'sass-rails',   '~> 3.2.3'
-  gem 'coffee-rails', '~> 3.2.1'
-  gem 'uglifier', '>= 1.0.3'
-  gem 'bootstrap-sass'
 end
