@@ -13,7 +13,7 @@
 //= require jquery
 //= require jquery_ujs
 //= require jquery.masonry.min.js
-//= require bootstrap
+//= require twitter/bootstrap
 //= require bootstrap-wysihtml5
 //= require bootstrap-wysihtml5/locales
 //= require_tree .
