@@ -8,8 +8,8 @@ Beeline::Application.routes.draw do
   get 'allproducts' => 'pages#allproducts'
   get 'comingsoon' => 'pages#comingsoon'
 
-  match 'contact' => 'contact#new', :as => 'contact', :via => :get
-  match 'contact' => 'contact#create', :as => 'contact', :via => :post
+  get 'contact' => 'contact#new'
+  post 'contact' => 'contact#create'
   
   root :to => 'products#index'
 
